move AGGREGATE_OPS to aggregate.ts
