+++ src/aggregate.ts
@@ -1,0 +1,8 @@
+export const AGGREGATE_OPS = [
+  'values', 'count', 'valid', 'missing', 'distinct',
+  'sum', 'mean', 'average', 'variance', 'variancep', 'stdev',
+  'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max',
+  'argmin', 'argmax'
+];
+
+// TODO: move supportedTypes, supportedEnums from schema to here
