+++ src/consts.ts
@@ -1,11 +1,3 @@
-// TODO: see if there is a nice way to import from Vega schema
-export const AGGREGATE_OPS = [
-  'values', 'count', 'valid', 'missing', 'distinct',
-  'sum', 'mean', 'average', 'variance', 'variancep', 'stdev',
-  'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max',
-  'argmin', 'argmax'
-];
-
 // TODO: remove
 export const MAXBINS_DEFAULT = 15;
